fix(clickhouse): los pesos de evidencia son con signo, y la columna no lo era (#33)
Los incidentes investigados no se estaban guardando. Ninguno.

evidence_weights era Array(UInt8), pero los pesos son con signo por diseño: la
ausencia de parecido visual vale -10, porque un detector que solo puede
escalar termina gritando lobo. ClickHouse rechazaba la fila entera con
"Unsigned type must not contain '-' symbol" y se perdía el incidente
completo — no solo ese peso.

Lo peor es a quién afectaba: exactamente a los incidentes investigados, que
son los que llevan captura y explicación, los que sostienen el argumento del
producto. Una corrida limpia guardaba 20 incidentes con 0 capturas y 0
explicaciones, y el error quedaba enterrado entre los logs del demo.

Columna migrada a Array(Int16). Tras el arreglo, la misma corrida guarda 20
incidentes, 20 con captura y 13 con explicación, sin fallos de inserción.

Verificado además: build limpio, 21/21 tests, auditoría de idioma sin
pendientes, zero-egress PASS.

+++ infra/clickhouse/schema.sql
@@ -71,7 +71,12 @@ CREATE TABLE IF NOT EXISTS sentinel.dns_incidents
     source_hosts         Array(String),
     domains              Array(String),
     evidence_types       Array(String),
-    evidence_weights     Array(UInt8),
+    -- Int16, not UInt8: evidence weights are signed by design. The absence of
+    -- visual similarity carries -10, because a detector that can only escalate
+    -- eventually cries wolf. With an unsigned column the whole insert failed
+    -- with "Unsigned type must not contain '-' symbol", which silently dropped
+    -- exactly the investigated incidents — the ones that matter most.
+    evidence_weights     Array(Int16),
     evidence_descriptions Array(String),
     explanation          String,
     recommended_action   String,
